requirements.txt
requests
twilio
python-dotenv
google-genai
pinecone
PyPDF2
sentence_transformers
pinecone